TNO-2365 Move email send to service manager
