+++ services/net/notification/NotificationManager.cs
@@ -44,16 +44,6 @@ public class NotificationManager : ServiceManager<NotificationOptions>
     protected INotificationEngine NotificationEngine { get; }
 
     /// <summary>
-    /// get - CHES service.
-    /// </summary>
-    protected IChesService Ches { get; }
-
-    /// <summary>
-    /// get - CHES options.
-    /// </summary>
-    protected ChesOptions ChesOptions { get; }
-
-    /// <summary>
     /// get - Template options.
     /// </summary>
     protected TemplateOptions TemplateOptions { get; }
@@ -91,12 +81,10 @@ public class NotificationManager : ServiceManager<NotificationOptions>
         IOptions<TemplateOptions> reportingOptions,
         INotificationValidator notificationValidator,
         ILogger<NotificationManager> logger)
-        : base(api, notificationOptions, logger)
+        : base(api, chesService, chesOptions, notificationOptions, logger)
     {
         _user = user;
         this.NotificationEngine = notificationEngine;
-        this.Ches = chesService;
-        this.ChesOptions = chesOptions.Value;
         this.TemplateOptions = reportingOptions.Value;
         _serializationOptions = serializationOptions.Value;
         this.NotificationValidator = notificationValidator as NotificationValidator ?? throw new ArgumentException("NotificationValidator must be of the correct type");
@@ -108,39 +96,6 @@ public class NotificationManager : ServiceManager<NotificationOptions>
     #endregion
 
     #region Methods
-    /// <summary>
-    /// Send email alert of failure.
-    /// </summary>
-    /// <param name="subject"></param>
-    /// <param name="message"></param>
-    /// <returns></returns>
-    public async Task SendEmailAsync(string subject, string message)
-    {
-        if (this.Options.SendEmailOnFailure)
-        {
-            try
-            {
-                var email = new TNO.Ches.Models.EmailModel(this.ChesOptions.From, this.Options.EmailTo, subject, message);
-                await this.Ches.SendEmailAsync(email);
-            }
-            catch (Exception ex)
-            {
-                this.Logger.LogError(ex, "Email failed to send");
-            }
-        }
-    }
-
-    /// <summary>
-    /// Send email alert of failure.
-    /// </summary>
-    /// <param name="subject"></param>
-    /// <param name="ex"></param>
-    /// <returns></returns>
-    public async Task SendEmailAsync(string subject, Exception ex)
-    {
-        await this.SendEmailAsync($"Notification Service - {subject}", $"<div>{ex.Message}</div>{Environment.NewLine}<div>{ex.StackTrace}</div>");
-    }
-
     /// <summary>
     /// Listen to active topics and import content.
     /// </summary>
